A lot of changes to the whole layout, dia and ignore file
The ignore file has a couple of new lines for private and build files.
BlockDevice has removed. It didn't have any real functionality.
MemoryBlockDevice has been renamed to FileSystem as it better represents its
functionality and responsibilities. It has been changed from a 'dumb'
container class to being responsible for creating files and folders, reading
and writing files in form of byte arrays and keeping track of the working
directory.

FolderBlock is a new class which keep tracks of files and their name. It
uses the standard block form of a byte array to save its data. This class
isn't finished.

Inode has the metadata of file, as its size, and also keep tracks of its
first data block.

Some small changes has been done to Shell and TestShell to take be able to
ahndle the changes int he other classes.
	modified:   .gitignore
	deleted:    Given kod/BlockDevice.java
	renamed:    Given kod/Filesystem.java -> Given kod/FileManager.java
	modified:   Given kod/Filesystem.java
	new file:   Given kod/FolderBlock.java
	new file:   Given kod/Inode.java
	deleted:    Given kod/MemoryBlockDevice.java
	modified:   Given kod/Shell.java
	modified:   Given kod/TestShell.java
	modified:   Layout.dia
	modified:   nbproject/project.properties