Changed implementation of the FolderBlock class.
The save() and load() are now static methods and uses a byte array which is
written with a serialized folderBlock. This data is turned back to an
instance with load(). using serialization forced me to change them to static
methods. The internal structure is now a map containing pairs of integer
fileId's and the filenames.